+++ Lib/Utils/CommandLineParser/Definitions/BuildInteractiveCommand.cs
@@ -10,6 +10,8 @@ namespace Lib.Utils.CommandLineParser.Definitions
 
         public CommandLineArgumentString Port { get; } = new CommandLineArgumentString(description: "set port for server to listen to", words: new[] { "-p", "--port" }, defaultValue: "8080");
 
+        public CommandLineArgumentBool Sprite { get; } = new CommandLineArgumentBool(description: "enable/disable creation of sprites", words: new[] { "--sprite" }, defaultValue: false);
+
         public CommandLineArgumentSwitch Verbose { get; } = new CommandLineArgumentSwitch(description: "enable spamming console output", words: new[] { "--verbose" });
     }
 }